# Optimized instance creation via activator