@jkotas 2019-04-16
>  An application fails with a NullReferenceException and I can't understand why it happens

My guess is that the function pointers are not getting setup correctly in your change.

> here is a little performance degradation in the class initialization case

I see up to ~20% performance degradation. E.g. 
```
    for (int i = 0; i < 100000000; i++)
        GC.KeepAlive(Activator.CreateInstance<object>());
```
6593ms before, 5485ms after.