@YohDeadfall 2019-04-17
> My guess is that the function pointers are not getting setup correctly in your change.

How can this be fixed? I use function pointers only to build delegates to constructors like it's done now in the `RuntimeType` class.