@YohDeadfall 2019-04-24
```
BenchmarkDotNet=v0.11.5, OS=Windows 10.0.18362
Intel Xeon CPU E5-2687W 0 3.10GHz, 2 CPU, 32 logical and 16 physical cores
.NET Core SDK=3.0.100-preview5-011517
  [Host] : .NET Core ? (CoreCLR 4.6.27717.0, CoreFX 4.700.19.21910), 64bit RyuJIT

Job=InProcess  Toolchain=InProcessEmitToolchain
```

**Before**

|       Method |     Mean |    Error |    StdDev |        Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------- |---------:|---------:|----------:|------------:|-------:|------:|------:|----------:|
|  CreateClass | 155.4 ns | 3.752 ns | 11.063 ns | 6,436,987.2 | 0.0062 |     - |     - |      40 B |
| CreateStruct | 148.7 ns | 2.957 ns |  4.048 ns | 6,726,591.1 | 0.0050 |     - |     - |      32 B |

**After**

|       Method |      Mean |     Error |    StdDev |         Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------- |----------:|----------:|----------:|-------------:|-------:|------:|------:|----------:|
|  CreateClass | 163.57 ns | 5.6100 ns | 16.453 ns |  6,113,460.1 | 0.0062 |     - |     - |      40 B |
| CreateStruct |  16.61 ns | 0.5319 ns |  1.568 ns | 60,208,064.5 |      - |     - |     - |         - |