@YohDeadfall 2019-04-24
@jkotas How can I get codegen of the `CreateInstance` method? Set `COMPlus_JitDisasm` to `CreateInstance`, but there is nothing printed to stdout.

```csharp
static void Main(string[] args) => GC.KeepAlive(Activator.CreateInstance<Address>());
```