@jkotas 2019-04-24
My favorite method is to attach debugger and look at the disassembly in the debugger.