@YohDeadfall 2019-05-03
@jkotas I've optimized my code by eliminating multiple static field access. So even `CreateInstance<T>` for reference types works faster now. Ready for the next review.

```
BenchmarkDotNet=v0.11.5, OS=Windows 10.0.18885
Intel Xeon CPU E5-2687W 0 3.10GHz, 2 CPU, 32 logical and 16 physical cores
.NET Core SDK=3.0.100-preview3-010431
  [Host] : .NET Core ? (CoreCLR 4.6.27717.0, CoreFX 4.700.19.22902), 64bit RyuJIT

Job=InProcess  Toolchain=InProcessEmitToolchain
```

**Before**

|       Method |     Mean |    Error |   StdDev |   Median |        Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------- |---------:|---------:|---------:|---------:|------------:|-------:|------:|------:|----------:|
|  CreateClass | 150.5 ns | 4.635 ns | 13.67 ns | 153.8 ns | 6,646,699.9 | 0.0062 |     - |     - |      40 B |
| CreateStruct | 133.6 ns | 5.106 ns | 15.06 ns | 139.8 ns | 7,485,729.8 | 0.0050 |     - |     - |      32 B |

**After**

|       Method |      Mean |     Error |     StdDev |         Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------- |----------:|----------:|-----------:|-------------:|-------:|------:|------:|----------:|
|  CreateClass | 137.59 ns | 3.8041 ns | 11.2165 ns |  7,268,009.4 | 0.0062 |     - |     - |      40 B |
| CreateStruct |  17.20 ns | 0.4269 ns |  0.9371 ns | 58,133,713.6 |      - |     - |     - |         - |