@jkotas 2019-05-05
@dotnet-bot test Windows_NT x64 Checked CoreFX Tests please