@roji 2019-05-06
@dotnet-bot test Windows_NT x64 Checked CoreFX Tests please