@roji 2019-05-11
@dotnet-bot test Windows_NT x64 Checked CoreFX Tests please