@karelz 2019-05-21
@jkotas ping?