@jkotas 2019-05-21
Does this implementation have good enough performance to replace Reflection.Emit workarounds like https://github.com/dotnet/corefx/blob/master/src/System.Text.Json/src/System/Text/Json/Serialization/ReflectionEmitMaterializer.cs ?