@YohDeadfall 2019-05-21
The following code doesn't box structs like the serializer does. Even if it won't be used by the serializer, this change fixes #16621.

```csharp
[Benchmark]
public Address ActivatorCreateClass() => Activator.CreateInstance<Address>();

[Benchmark]
public Coordinates ActivatorCreateStruct() => Activator.CreateInstance<Coordinates>();

private static readonly Func<Address> s_addressCtor = CreateConstructor<Address>();
private static readonly Func<Coordinates> s_coordinatesCtor = CreateConstructor<Coordinates>();

[Benchmark]
public Address EmitCreateClass() => s_addressCtor();

[Benchmark]
public Coordinates EmitCreateStruct() => s_coordinatesCtor();

private static Func<T> CreateConstructor<T>()

{
    Type type = typeof(T);
    ConstructorInfo realMethod = type.GetConstructor(BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance, binder: null, Type.EmptyTypes, modifiers: null);

    if (realMethod == null && !type.IsValueType)
    {
        return null;
    }

    var dynamicMethod = new DynamicMethod(
        ConstructorInfo.ConstructorName,
        typeof(T),
        Type.EmptyTypes,
        typeof(ActivatorTests).Module,
        skipVisibility: true);

    ILGenerator generator = dynamicMethod.GetILGenerator();

    if (realMethod == null)
    {
        LocalBuilder local = generator.DeclareLocal(type);

        generator.Emit(OpCodes.Ldloca_S, local);
        generator.Emit(OpCodes.Initobj, type);
        generator.Emit(OpCodes.Ldloc, local);
    }
    else
    {
        generator.Emit(OpCodes.Newobj, realMethod);
    }

    generator.Emit(OpCodes.Ret);

    return (Func<T>)dynamicMethod.CreateDelegate(typeof(Func<T>));
}
```

```ini
BenchmarkDotNet=v0.11.5, OS=Windows 10.0.18895
Intel Xeon CPU E5-2687W 0 3.10GHz, 1 CPU, 16 logical and 16 physical cores
.NET Core SDK=3.0.100-preview5-011568
  [Host] : .NET Core ? (CoreCLR 4.6.27729.0, CoreFX 4.700.19.22902), 64bit RyuJIT

Job=InProcess  Toolchain=InProcessEmitToolchain
```

|                Method |       Mean |     Error |     StdDev |     Median |          Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------------- |-----------:|----------:|-----------:|-----------:|--------------:|-------:|------:|------:|----------:|
|  ActivatorCreateClass | 139.801 ns | 4.0067 ns | 11.8138 ns | 142.920 ns |   7,152,999.3 | 0.0062 |     - |     - |      40 B |
| ActivatorCreateStruct |  12.712 ns | 0.6657 ns |  1.9630 ns |  13.374 ns |  78,667,869.9 |      - |     - |     - |         - |
|       EmitCreateClass |  23.641 ns | 1.5155 ns |  4.4685 ns |  25.585 ns |  42,299,944.8 | 0.0063 |     - |     - |      40 B |
|      EmitCreateStruct |   6.413 ns | 0.2474 ns |  0.7295 ns |   6.559 ns | 155,932,503.6 |      - |     - |     - |         - |

/cc @ahsonkhan @steveharter @xoofx