@jkotas 2019-06-11
The change makes the startup cost of `Activator<T>` up to 6x more expensive. 

E.g. https://gist.github.com/jkotas/421b3c44099347eb9a18889c2cf9e00e runs in 1s before this change, and 6s with this change.

I think it is expected that this optimization will have some startup cost hit with this implementation strategy, but it would be nice to make the regression as small as possible. Can we do better than 6x?