@YohDeadfall 2019-06-12
Probably we can, because there is `GetDefaultConstructor` in `RuntimeTypeHandle` and no binder is used. Will take a try tomorrow.